Refactor HeroPage data fetching
Move data fetching logic for HeroPage to _getPage function and use withPageData HOC for fetching page data. Removed console.log statements and commented-out legacy code for cleaner implementation.
