+++ frontend-next-migration/src/preparedPages/HeroesPages/ui/HeroPage.tsx
@@ -33,6 +33,9 @@ const HeroPage: FC<Props> = ({
   prevHeroLink,
   nextHeroLink,
 }) => {
+
+
+
   return (
     <main className={cls.main}>
       <HeroContainer
